Moved user db into @example/database
