+++ README.md
@@ -1,15 +1,20 @@
 # Paima Engine
 
+# Development
+
 ## Install dependencies
 `deno install --allow-scripts`
 
 ## Run Example Deployment
 `deno task -f @example/node dev`
 
-## Run Example Webpage
-> NOTE: Example Deployment must be running
+## Create pgtypes
+`deno task -f @example/database pgtyped:update`
 
+## Launch Explorer
 `deno task -f @example/frontend start`
 
+# Tests
 # Run Tests
 `deno task -f @example/node test`
+
