+++ deno.lock
@@ -5506,7 +5506,7 @@
   },
   "workspace": {
     "members": {
-      "example/backend/state-transition/src": {
+      "example/backend/database": {
         "dependencies": [
           "npm:@pgtyped/runtime@2.4.2"
         ]
