+++ example/backend/database/deno.json
@@ -1,0 +1,12 @@
+{
+    "name": "@example/database",
+    "version": "0.3.0",
+    "exports": "./src/mod.ts",
+    "imports": {
+        "@pgtyped/runtime": "npm:@pgtyped/runtime@2.4.2"
+    },
+    "tasks": {
+        "pgtyped:my-sql": "deno task -f @paima/db db:wait && pgtyped -w -c ./pgtypedconfig.json",
+        "pgtyped:update": "MIGRATIONS=\"`pwd`/src/migrations/\" concurrently --raw --kill-others \"deno task -f @paima/db db:up\" \"deno task -f @paima/db pgtyped:internal\" \"deno task pgtyped:my-sql\""
+    }
+}
