+++ example/backend/database/pgtypedconfig.json
@@ -11,7 +11,7 @@
       "emitTemplate": "{{dir}}/{{name}}.types.ts"
     }
   ],
-  "srcDir": "./sql",
+  "srcDir": "./src/sql",
   "failOnError": false,
   "camelCaseColumnNames": false,
   "db": {
