+++ example/backend/database/src/mod.ts
@@ -1,0 +1,2 @@
+export { insertStateMachineInput } from "./sql/sm_example.queries.ts";
+export { migrations } from "./migrationts.ts";
