+++ example/backend/state-transition/deno.json
@@ -1,0 +1,8 @@
+{
+  "name": "@example/state-transition",
+  "version": "0.3.0",
+  "exports": "./src/mod.ts",
+  "tasks": {
+    "my-files": "deno task -f @paima/db db:wait && pgtyped -w -c ./pgtypedconfig.json",
+  }
+}
