+++ example/backend/state-transition/src/mod.ts
@@ -1,3 +1,1 @@
 export * from "./state-machine.ts";
-export * from "./sql/sm_example.queries.ts";
-export * from "./migrationts.ts";
